+++ angular-view/src/app/app.module.ts
@@ -1,8 +1,9 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule } from '@angular/forms';
-
+import { HttpClientModule } from "@angular/common/http";
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+
 import { MatButtonModule } from '@angular/material/button';
 import { MatInputModule } from '@angular/material/input';
 import { MatMenuModule } from '@angular/material/menu';
@@ -10,6 +11,7 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatCardModule } from '@angular/material/card';
 import { MatListModule } from '@angular/material/list';
 import { MatDialogModule } from '@angular/material/dialog';
+import { MatFormFieldModule } from '@angular/material/form-field';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
@@ -21,6 +23,17 @@ import { ButtonMenuSelectComponent } from './components/button-menu-select/butto
 
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
 import { AuthInterceptor } from './interceptors/auth-interceptor';
+
+const matModules = [
+  MatButtonModule,
+  MatInputModule,
+  MatMenuModule,
+  MatIconModule,
+  MatCardModule,
+  MatListModule,
+  MatDialogModule,
+  MatFormFieldModule,
+];
 
 @NgModule({
   declarations: [
@@ -35,14 +48,9 @@ import { AuthInterceptor } from './interceptors/auth-interceptor';
     BrowserModule,
     AppRoutingModule,
     FormsModule,
+    HttpClientModule,
     BrowserAnimationsModule,
-    MatButtonModule,
-    MatInputModule,
-    MatMenuModule,
-    MatIconModule,
-    MatCardModule,
-    MatListModule,
-    MatDialogModule,
+    ...matModules
   ],
   providers: [
     {