+++ angular-view/src/app/service/message.service.ts
@@ -1,0 +1,11 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+    providedIn: 'root'
+})
+export class MessageService {
+
+    add(message: string) {
+        // TODO: show message at user gui
+    }
+}