+++ express-server/src/database/user.model.ts
@@ -2,7 +2,7 @@ import * as mongoose from 'mongoose';
 
 const UserSchema = new mongoose.Schema({
     googleId: { type: String, },
-    name: { type: String },
+    displayName: { type: String },
     email: {
         type: String,
         match: /[a-z0–9!#$%&’*+/=?^_`{|}~-]+(?:\.[a-z0–9!#$%&’*+/=?^_`{|}~-]+)*@(?:[a-z0–9](?:[a-z0–9-]*[a-z0–9])?\.)+[a-z0–9](?:[a-z0–9-]*[a-z0–9])?/,